+++ R/eq5d.R
@@ -1,7 +1,7 @@
 #' Calculate EQ-5D index scores
 #'
-#' Wrapper for \code{eq5d3l}, \code{eq5d5l} and \code{eq5dy}. Calculate EQ-5D index scores for
-#' EQ-5D-3L, EQ-5D-5L and EQ-5D-Y. Available value sets can be viewed using the function
+#' Wrapper for \code{eq5d3l}, \code{eq5d5l} and \code{eq5dy3l}. Calculate EQ-5D index scores for
+#' EQ-5D-3L, EQ-5D-5L and EQ-5D-Y-3L. Available value sets can be viewed using the function
 #' \code{valuesets}.
 #'
 #' @param scores numeric or data.frame with names/colnames MO, SC, UA, PD and AD
@@ -9,7 +9,7 @@
 #'   Anxiety/depression. Alternatively EQ-5D scores can be provided in
 #'   five digit format e.g. 12321. If five digit scores are used in a data.frame
 #'   the default column name look for by the function is "State".
-#' @param version string of value "3L", "5L" or "Y" to indicate instrument
+#' @param version string of value "3L", "5L" or "Y3L" to indicate instrument
 #'   version.
 #' @param type string specifying method type used in deriving value set scores.
 #'   Options are TTO or VAS for EQ-5D-3L, VT for EQ-5D-5L, CW for EQ-5D-5L
@@ -25,7 +25,8 @@
 #'   "Utility", "Age", "Sex" and "bwidth") for NICE DSU mapping. bwidth can also
 #'   be a number which is applied to the whole dataset. When a single
 #'   NICE DSU score is being calculated "age", "sex" and "bwidth" are also
-#'   used. See \code{\link{eq5dmap}} for valid options.
+#'   used. See \code{\link{eq5dmap}} for valid options. "digits" can also be 
+#'   used to return NICE DSU mapping scores with more precision.
 #' @return a numeric vector of utility index scores.
 #' @examples
 #' 
@@ -91,6 +92,7 @@ eq5d.data.frame <- function(scores, version=NULL, type=NULL, country=NULL, ignor
   sex <- "Sex"
   age <- "Age"
   bwidth <- "bwidth"
+  digits <- 3
 
   if(!is.null(args$dimensions)) {dimensions <- args$dimensions}
   if(!is.null(args$five.digit)) {five.digit <- args$five.digit}
@@ -98,6 +100,7 @@ eq5d.data.frame <- function(scores, version=NULL, type=NULL, country=NULL, ignor
   if(!is.null(args$sex)) {sex <- args$sex}
   if(!is.null(args$age)) {age <- args$age}
   if(!is.null(args$bwidth)) {bwidth <- args$bwidth}
+  if(!is.null(args$digits)) {digits <- args$digits}
 
   eq5d.columns <- NULL
   if(all(dimensions %in% names(scores))) {
@@ -122,11 +125,11 @@ eq5d.data.frame <- function(scores, version=NULL, type=NULL, country=NULL, ignor
   res <- apply(scores, 1, function(x) {
     if(type=="DSU") {
       if(is.null(bwidth)) {
-        eq5d.default(x[eq5d.columns], version=version, type=type, country=country, ignore.invalid=ignore.invalid, age=x[age], sex=x[sex])
+        eq5d.default(x[eq5d.columns], version=version, type=type, country=country, ignore.invalid=ignore.invalid, age=x[age], sex=x[sex], digits=digits)
       } else if(is.numeric(bwidth)) {
-        eq5d.default(x[eq5d.columns], version=version, type=type, country=country, ignore.invalid=ignore.invalid, age=x[age], sex=x[sex], bwidth=bwidth)
+        eq5d.default(x[eq5d.columns], version=version, type=type, country=country, ignore.invalid=ignore.invalid, age=x[age], sex=x[sex], bwidth=bwidth, digits=digits)
       } else {
-        eq5d.default(x[eq5d.columns], version=version, type=type, country=country, ignore.invalid=ignore.invalid, age=x[age], sex=x[sex], bwidth=x[bwidth])
+        eq5d.default(x[eq5d.columns], version=version, type=type, country=country, ignore.invalid=ignore.invalid, age=x[age], sex=x[sex], bwidth=x[bwidth], digits=digits)
       }
     } else {
       eq5d.default(x[eq5d.columns], version=version, type=type, country=country, ignore.invalid=ignore.invalid, ...)
@@ -143,9 +146,14 @@ eq5d.matrix <- function(scores, version=NULL, type=NULL, country=NULL, ignore.in
 
 #' @export
 eq5d.default <- function(scores, version=NULL, type=NULL, country=NULL, ignore.invalid=FALSE, ...){
-
-  if(!version %in% c("3L", "5L", "Y"))
-    stop("EQ-5D version not one of 3L, 5L or Y.")
+  
+  if (!is.null(version) && version == "Y") {
+    lifecycle::deprecate_warn("0.16.0", I('Setting `version = "Y"`'), I('`version = "Y3L"`'))
+    version <- "Y3L"
+  }
+  
+  if(!version %in% c("3L", "5L", "Y3L"))
+    stop("EQ-5D version not one of 3L, 5L or Y3L.")
 
   .length = length(scores)
   if(!is.null(type) && type=="DSU") {
@@ -212,6 +220,19 @@ eq5d.default <- function(scores, version=NULL, type=NULL, country=NULL, ignore.i
       stop("Sex must be Male, Female, M or F (case insensitive).")
     }
   }
+  
+  if(!is.null(args$digits)) {
+    digits <- suppressWarnings(as.numeric(args$digits))
+    if(is.na(digits)) {
+      if(ignore.invalid) {
+        return(NA)
+      } else {
+        stop("\"digits\" must be a numeric.")
+      }
+    }
+  } else {
+    digits <- 3
+  }
 
   .length <- length(scores)
   if(.length==5 && any(!scores %in% 1:.get_number_levels(version))) { #if length==5
@@ -238,12 +259,12 @@ eq5d.default <- function(scores, version=NULL, type=NULL, country=NULL, ignore.i
     } else if(!is.null(type) && type=="RCW") {
       eq5drcw(scores, country=country)
     } else if(!is.null(type) && type=="DSU") {
-      eq5dmap(scores, country, version, args$age, args$sex, bwidth)
+      eq5dmap(scores, country, version, args$age, args$sex, bwidth, digits)
     } else {
       stop("EQ-5D-3L valueset type not recognised. Must be one of 'TTO', 'VAS', 'RCW' or 'DSU'.")
     }
-  } else if (version=="Y") {
-    eq5dy(scores, country=country)
+  } else if (version=="Y3L") {
+    eq5dy3l(scores, country=country)
   }
   else {
     if(!is.null(type) && type=="VT") {
@@ -251,7 +272,7 @@ eq5d.default <- function(scores, version=NULL, type=NULL, country=NULL, ignore.i
     } else if(!is.null(type) && type=="CW") {
       eq5dcw(scores, country=country)
     } else if(!is.null(type) && type=="DSU") {
-      eq5dmap(scores, country, version, as.numeric(args$age), args$sex, bwidth)
+      eq5dmap(scores, country, version, as.numeric(args$age), args$sex, bwidth, digits)
     } else {
       stop("EQ-5D-5L valueset type not recognised. Must be one of 'VT', 'CW' or 'DSU'.")
     }
@@ -265,7 +286,7 @@ eq5d.default <- function(scores, version=NULL, type=NULL, country=NULL, ignore.i
 #'     in the \code{eq5d} package.
 #'
 #' @param type string EQ-5D value set type. TTO or VAS for EQ-5D-3L, VT for EQ-5D-5L,
-#'   cTTO for EQ-5D-Y, CW for EQ-5D-5L crosswalk conversion dataset, or DSU for NICE Decision Support
+#'   cTTO for EQ-5D-Y-3L, CW for EQ-5D-5L crosswalk conversion dataset, or DSU for NICE Decision Support
 #'   Unit's EQ-5D-5L to EQ-5D-3L and EQ-5D-3L to EQ-5D-5L mappings.
 #' @param version string either 3L, 5L or Y.
 #' @param country string one of the countries for which there is a value set.
@@ -280,9 +301,16 @@ eq5d.default <- function(scores, version=NULL, type=NULL, country=NULL, ignore.i
 #' valuesets(type="TTO")
 #' valuesets(version="5L")
 #' valuesets(country="UK")
-#' valuesets(version="Y", references=c("DOI", "PubMed"))
+#' valuesets(version="Y3L", references=c("DOI", "PubMed"))
 #' @export
 valuesets <- function(type=NULL, version=NULL, country=NULL, references=c("PubMed", "DOI", "ISBN", "ExternalURL")) {
+  
+  if (!is.null(version) && version == "Y") {
+    lifecycle::deprecate_warn("0.16.0", I('Setting `version = "Y"`'), I('`version = "Y3L"`'))
+    version <- "Y3L"
+  }
+  
+  if(!is.null(version) && version == "Y3L") version <- "Y-3L"
   if(!is.null(version)) version <- paste0("EQ-5D-", version)
 
   tto <- data.frame(Version="EQ-5D-3L", Type="TTO", Country=colnames(TTO))
@@ -290,7 +318,7 @@ valuesets <- function(type=NULL, version=NULL, country=NULL, references=c("PubMe
   rcw <- data.frame(Version="EQ-5D-3L", Type="RCW", Country=colnames(RCW))
   vt <- data.frame(Version="EQ-5D-5L", Type="VT", Country=colnames(VT))
   cw <- data.frame(Version="EQ-5D-5L", Type="CW", Country=colnames(CW))
-  y <- data.frame(Version="EQ-5D-Y", Type="cTTO", Country=colnames(Y))
+  y <- data.frame(Version="EQ-5D-Y-3L", Type="cTTO", Country=colnames(Y3L))
   dsu3l <- data.frame(Version="EQ-5D-3L", Type="DSU", Country=sub("Copula", "", grep("Copula", sort(colnames(DSU3L)), value=TRUE)))
   dsu5l <- data.frame(Version="EQ-5D-5L", Type="DSU", Country=sub("Copula", "", grep("Copula", sort(colnames(DSU5L)), value=TRUE)))
   vs <- rbind(tto, vas, rcw, vt, cw, y, dsu3l, dsu5l)